+++ tablefill/__init__.py
@@ -3,6 +3,6 @@
 
 __author__  = """Mauricio Cáceres Bravo"""
 __email__   = 'caceres@nber.org'
-__version__ = '0.9.5'
+__version__ = '0.9.6'
 
 from .tablefill import tablefill